requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Data Ingestion & Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0
apache-beam[gcp]==2.52.0
pyspark==3.5.0

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.1
google-cloud-bigquery==3.14.1
deltalake==0.14.0

# Feature Engineering
feast==0.36.0
tfx==1.14.0

# Machine Learning
torch==2.1.2
tensorflow==2.15.0
transformers==4.36.2
xgboost==2.0.3
lightgbm==4.2.0

# Time Series
prophet==1.1.5
statsmodels==0.14.1
tslearn==0.6.2

# Vector Search
faiss-cpu==1.7.4
qdrant-client==1.7.0

# MLOps & Monitoring
mlflow==2.9.2
wandb==0.16.2
dvc==3.38.1
evidently==0.4.13

# API & Web
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
plotly==5.18.0
dash==2.14.2

# RAG & GenAI
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
chromadb==0.4.22
sentence-transformers==2.2.2

# Explainability
shap==0.44.0
lime==0.2.0.1

# Deployment & Orchestration
docker==7.0.0
kubernetes==29.0.0
ray[serve]==2.9.0
celery==5.3.4
redis==5.0.1

# Testing & Quality
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring & Observability
prometheus-client==0.19.0
grafana-api==1.0.3
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Privacy & Security
cryptography==41.0.7
differential-privacy==1.1.1